# docs/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stoneboyz CRM API
  version: 0.1.0
  description: |
    Source of truth for all API contracts.
    Edit this file before implementing any endpoint.
    packages/api-client is generated from this file; do not hand-edit it.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.stoneboyz.app/v1
    description: Production placeholder
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QuoteId:
      name: quoteId
      in: path
      required: true
      description: UUID of the quote
      schema:
        type: string
        format: uuid
...
    AreaId:
      name: areaId
      in: path
      required: true
      description: UUID of the quote area
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: object
...
    GeneratedPriceLine:
      type: object
      required:
        - id
        - quoteAreaId
        - category
        - label
        - quantity
        - unit
        - unitPriceCents
        - lineTotalCents
        - priceListItemId
        - sortOrder
        - overridePriceCents
        - overrideReason
        - overrideByUserId
        - overrideAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        quoteAreaId:
          type: string
          format: uuid
        category:
          type: string
          enum: [material, fabrication, finished_edge, splash, sink_cutout, sink_item, faucet_hole]
        label:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        unitPriceCents:
          type: integer
          minimum: 0
        lineTotalCents:
          type: integer
          minimum: 0
        priceListItemId:
          type: string
          format: uuid
          nullable: true
        sortOrder:
          type: integer
        overridePriceCents:
          type: integer
          minimum: 0
          nullable: true
        overrideReason:
          type: string
          nullable: true
        overrideByUserId:
          type: string
          nullable: true
        overrideAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OverrideGeneratedPriceLineBody:
      type: object
      properties:
        overridePriceCents:
          type: integer
          minimum: 0
          nullable: true
        overrideReason:
          type: string
          minLength: 1
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/pricing:
    get:
      operationId: listQuoteAreaPricingLines
      summary: List generated pricing lines for a quote area
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Generated pricing lines returned.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeneratedPriceLine'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/pricing/generate:
    post:
      operationId: generateQuoteAreaPricingLines
      summary: Generate pricing lines for a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      responses:
        '201':
          description: Generated pricing lines replaced.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeneratedPriceLine'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/pricing/{lineId}/override:
    patch:
      operationId: overrideQuoteAreaPricingLine
      summary: Override a generated pricing line
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - name: lineId
          in: path
          required: true
          description: UUID of the generated price line
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverrideGeneratedPriceLineBody'
      responses:
        '200':
          description: Generated pricing line override updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedPriceLine'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generated price line, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...